\name{print.lin_reg}
\alias{print.lin_reg}
\title{Print an unitary linear regression}
\usage{
\method{print}{lin_reg}(model)
}
\arguments{
\item{model, }{fitted lin_reg model}
}
\description{
print.lin_reg is to print an unitary linear regression model
}
\details{
Just like the famous print.lm,
print.lin_reg is to print the information of an unitary linear regression,
including the author, intercept and the slope.
}
