% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simp_lin_R}
\alias{simp_lin_R}
\title{Fast Simple Linear Regression}
\usage{
simp_lin_R(x, y)
}
\arguments{
\item{x}{A numeric vector of values}
\item{y}{A numeric vector of values with same length as x}
}
\description{Performs simple linear regression with x as the predictor and y as the response. Function is written in Rcpp for improved computational efficiency.
}
\examples{
x <- rnorm(1e2)
y = x + rnorm(1e2)
simp_lin_R(x, y)
}
\keyword{ function }
\author{Justin Van Ee (vanee002@colostate.edu)}
