\name{SimpLin-package}
\alias{SimpLin-package}
\alias{SimpLin}
\docType{package}
\title{
  Simple linear regression
}
\description{
  The package contains a single function simp_lin_R().
}

\author{
Seongwon Im, rseongwoni1@gmail.com

Maintainer: Seongwon Im <rseongwoni1@gmail.com>
}
\references{
  Kutner, M. H., Nachtsheim, C. J., Neter, J., &amp; Li, W. (2013). Applied linear statistical models (5th ed.). New York, NY: McGraw-Hill Education Private Limited.
}
\keyword{ package }