% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimpLinCpp}
\alias{SimpLinCpp}
\title{Cpp function to obtain
  estimated regression coefficients
  standard errors of the estimates
  95% confidence interval of each coefficient
  predicted values
  residuals}
\usage{
SimpLinCpp(x, y, n, sig_level)
}
\arguments{
\item{x}{predictor, n by 1 numeric vector}

\item{y}{response, n by 1 numeric vector}

\item{n}{number of observations, integer}

\item{sig_level}{significance level of the confidence interval, double}
}
\value{
result of fit List
}
\description{
Cpp function to obtain
  estimated regression coefficients
  standard errors of the estimates
  95% confidence interval of each coefficient
  predicted values
  residuals
}
