% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLinR.R
\name{SimpLinR}
\alias{SimpLinR}
\title{Perform simple linear regression analysis.}
\usage{
SimpLinR(x, y)
}
\arguments{
\item{x}{A numeric vector of length n.}

\item{y}{A numeric vector of length n.}
}
\value{
A named list with five elements: estimated parameter coefficients, standard errors for these coefficients, 95\% confidence intervals for these coefficients, predicted values, and residuals.
}
\description{
Perform simple linear regression analysis.
}
\examples{
x <- c(1,5,6)
y <- c(1,2,3)
fit <- SimpLinR(x, y)
# estimated coefficients, with intercept always first
fit$coefs
# standard errors of both coefficients
fit$std_errors
# 95\% confidence interval for both coefficients; 2.5\% is the first column and 97.5\% is the second column
fit$conf_int
# predicted values
fit$preds
# residuals
fit$resids
}
