\name{summary.lin_reg}
\alias{summary.lin_reg}
\title{Summarizing unary linear regression model fits}
\usage{
\method{summary}{lin_reg}(model)
}
\arguments{
\item{model, }{unary linear model}
}
\value{
a list, more information about the model include created time
}
\description{
summary method for class "lin_reg".
}
\details{
Summary for lin_reg model.
}
