% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{simp_lin_R}
\alias{simp_lin_R}
\title{Wraps the Rcpp simple linear regression function and provides error checking.}
\usage{
simp_lin_R(x, y)
}
\arguments{
\item{obj}{The}
}
\value{
Product of v1 and v2
}
\description{
Wraps the Rcpp simple linear regression function and provides error checking.
}
