\name{simplin-package}
\alias{simplin-package}
\alias{simplin}
\docType{package}
\title{
\packageTitle{simplin}
}
\description{
Perform simple linear regression and return estimate of coefficients,predicted responses, residuals, standard errors of coefficients and 95 percents confidence intervals of coefficients.
}
\usage{
simp_lin_R()
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simplin}
\packageIndices{simplin}
Use simp_lin_R function to obtain a list of all the statistics
}
\author{
Bowen Li email:bowenli@colostate.edu
\packageAuthor{simplin}

Maintainer: \packageMaintainer{simplin}
}
\references{
%% ~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
