\name{SimpLin-package}
\alias{SimpLin-package}
\alias{SimpLin}
\docType{package}
\title{
\packageTitle{SimpLin}
}
\description{
\packageDescription{SimpLin}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SimpLin}
\packageIndices{SimpLin}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SimpLin}

Maintainer: \packageMaintainer{SimpLin}
}
\references{~~ Literature or other references for background information ~~}
\keyword{ package }

