\name{SimpLin-package}
\alias{SimpLin-package}
\alias{SimpLin}
\docType{package}
\title{
\packageTitle{SimpLin}
}
\description{
\packageDescription{SimpLin}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{SimpLin}

Maintainer: \packageMaintainer{SimpLin}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = SimpLin)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{SimpLin}}
%\section{Documentation Index}{\packageIndices{SimpLin}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
