\name{SimpLin-package}
\alias{SimpLin-package}
\alias{SimpLin}
\docType{package}
\title{
\packageTitle{SimpLin}
}
\description{
\packageDescription{SimpLin}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SimpLin}
\packageIndices{SimpLin}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SimpLin}

Maintainer: \packageMaintainer{SimpLin}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
