\name{simplin-package}
\alias{simplin-package}
\alias{simplin}
\docType{package}
\title{
  The R Simp_lin Package
}
\description{
R simple linear regression functions.
}
\details{
This package contains a function called simp_lin_R for fitting a simple linear regression model.
}
\author{
Xiangdong Meng

Maintainer: Xiangdong Meng <mengxd@colostate.edu>
}

\keyword{ package }

