\name{SimpLin-package}
\alias{SimpLin-package}
\alias{SimpLin}
\docType{package}
\title{\packageTitle{SimpLin}}
\description{\packageDescription{SimpLin}}
\section{Package Content}{\packageIndices{SimpLin}}
\author{\packageAuthor{SimpLin}}
\section{Maintainer}{\packageMaintainer{SimpLin}}
\keyword{package}
