% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLin_Wrapper.R
\name{SimpLin}
\alias{SimpLin}
\title{A function that fits a simple linear regression model}
\usage{
SimpLin(x, y)
}
\arguments{
\item{x}{A numeric vector serving as the explanatory variable.}

\item{y}{A numeric vector equal in length to x serving as the response
variable.}
}
\value{
A list. [\link{"coefficients"} will give the beta vector where the first
element is the intercept term and the second element is the slope term,
[\link{"SE B0"}] will give SE of the intercept, will give the beta vector,
[\link{"SE B1"}] will give SE of the slope, [\link{"95\% CI B0"}] will give a 95\%
confidence for the intercept, [\link{"95\% CI B1"}] will give a 95\% confidence
for the slope, [\link{"yhat}] will give the predicted values, [\link{"residuals"}]
will give the residuals.
}
\description{
This function takes two vectors, x and y, as inputs. It then fits a simple
regression model and returns a list with regression coefficients, their
corresponding standard errors and 95\% confidence intervals, residuals,
and predicted values as a list.
}
