% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLinR.R
\name{SimpLinR}
\alias{SimpLinR}
\title{Fit a simple linear regression model}
\usage{
SimpLinR(x, y, significance_level = 0.05)
}
\arguments{
\item{x}{A predictor variable, n by 1 numeric vector}

\item{y}{A response variable, n by 1 numeric vector}

\item{significance_level}{Significance level for confidence interval}
}
\value{
A list of the fit
}
\description{
This function fits a simple linear regression model using an n by 1
predictor variable (x) and an n by 1 response variable (y).
It throws an error if
   i) length of x and y is different
   ii) x is not numeric
   iii) y is not numeric
It returns a fitted result, which includes
   estimated regression coefficients
   standard errors of the estimates
   95% confidence interval of each coefficient
   predicted values
   residuals
}
\examples{
n <- 100
x <- rnorm(n)
beta0 <- 1
beta1 <- -1
y <- beta0 + beta1 * x + rnorm(n)
SimpLinR(x, y)
}
