% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLinR.R
\name{SimpLinR}
\alias{SimpLinR}
\title{Simple Linear Regression}
\usage{
SimpLinR(x, y)
}
\arguments{
\item{x}{Numeric vector containing the predictor variable}

\item{y}{Numeric vector containing the response variable}
}
\value{
A list containing five separate elements: a vector of coefficients, a vector of their standard errors, a matrix of the confidence intervals of the coefficients, a vector of the residuals, and a vector of the fitted values
}
\description{
This function takes in two numeric vectors, the predictor variable and the
the response variable, and runs a simple linear regression
}
