% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLinR.R
\name{SimpLinR}
\alias{SimpLinR}
\title{Simple Linear Regression (minimal, with SE & 95\% CI)}
\usage{
SimpLinR(x, y)
}
\arguments{
\item{x}{numeric vector (predictor)}

\item{y}{numeric vector (response)}
}
\value{
list: beta0_hat, beta1_hat, se_beta0, se_beta1,
ci_beta0_95, ci_beta1_95, residuals, fitted, predicted
}
\description{
Simple Linear Regression (minimal, with SE & 95\% CI)
}
