% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLinR.R
\name{SimpLinR}
\alias{SimpLinR}
\title{Simple Linear Regression}
\usage{
SimpLinR(x, y)
}
\arguments{
\item{x}{vector for independent variable}

\item{y}{vector for dependent variable}
}
\value{
List of different linear regression metrics: fitted coefficients, predictions, residuals, standard errors of coefficients, and 95% confidence interval of coefficients
}
\description{
Simple Linear Regression
}
\examples{
x <- c(1, 2, 3, 4)
y <- c(2, 4, 6, 11)
SimpLinR(x, y)

}
