% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLinR.R
\name{SimpLinR}
\alias{SimpLinR}
\title{Simple Linear Regression}
\usage{
SimpLinR(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
Simple Linear Regression
}
\examples{
x <- 1:100
y <- 2 * x + rnorm(100, mean = 0, sd = 20)
SimpLinR(x, y)
}
